Update website links to direct internally within the page rather than referring to the archived version.
The web.archive.org prefix causing a redirection to the internet archive was removed in the website's HTML file. Consequently, the updated links now lead to the desired webpage within the site itself. This prevents users from mistakenly navigating to an outdated, archived version of the page.

href="https://web.archive.org/web/20231207160314